Data migration to fix submitted/rejected state
The `Edition`s in this data migration have all been unpublished and then
submitted (and in some cases subsequently rejected). This results in
them 404-ing instead of redirecting in production currently and in them
not republishing correctly when being migrated.

There is no valid transition from these states to `draft` but setting
them to draft will cause all of the correct behaviour with no side
effects so... data migration.